==> Code/03_Construccion_BBDD_Temperaturas.ipynb <==
import numpy as np
import pandas as pd
from pandas import Series,DataFrame
import sqlite3

De todas las variables que tenemos disponibles, para hacer el primer análisis nos vamos a quedar con las siguientes

TMAX = Maximum temperature (tenths of degrees C)

TMIN = Minimum temperature (tenths of degrees C)

TAVG = Average temperature (tenths of degrees C)

PRCP = Precipitation (tenths of mm)

El resto de variables disponibles para su explotación, no están lo suficientemente informadas (o están sin informar)

import numpy as np
import pandas as pd
from pandas import Series,DataFrame
import sqlite3

# Una vez realizadas las pruebas necesarias para obtener el dataframe adecuado, procedemos a generar la base total
# creacion de la conexion
conn = sqlite3.connect("D:/master/Master-in-Data-Science/Proyecto/Datos/worldTimeSeries.sqlite")

# creacion del cursor
cursor = conn.cursor()

# Eliminacion de la tabla si existe previamente
res = cursor.execute("drop table if exists world_time;")

# Creacion de la tabla 
res = cursor.execute("""create table world_time (
                     I NUMERIC,
                     STATION TEXT,
                     DATE TEXT,
                     TMAX REAL,
                     TMIN REAL,
                     TAVG REAL,
                     PRCP REAL);
                     """)

#                     ,PRIMARY KEY (STATION, DATE));

# Se lee el fichero con la reducción de estaciones
cat = pd.read_csv('../datos/ghcnd-stations-cluster.csv',sep=";")

# Las estaciones en los archivos de datos tienen concatenado el código de pais y el codido de estacion, por tanto 
# creamos una nueva columna con ese formato
cat['STATION'] = cat['COUNTRY']+cat['ORIG_ID']
cat['NEW_STATION'] = cat['COUNTRY']+cat['ID']

# Nos quedamos con las columnas necesarias, estacion original y centroide
cat =cat[["STATION","NEW_STATION"]]

for year in range(1763, 2017):
#for year in range(2015, 2016):
    print("Ejecutando año: " + str(year))
    # Lectura del archivo de datos
    df = pd.read_csv('D:/master/Master-in-Data-Science/Proyecto/Datos/'+str(year)+'.csv.gz',header=None,
                    names=['STATION','DATE','MEASURE','VALUE','M-FLAG','Q-FLAG','S-FLAG','TIME'],
                    dtype={'DATE':str }
                    )
    
    # Nos quedamos con las columnas que nos interesan
    df=df[["STATION","DATE","MEASURE","VALUE"]]
    # Se eliminan las columnas que tienen valores missing
    df = df[df['VALUE'] != -999]
    
    # Se construye el DataFrame base
    df_final = DataFrame(df[['STATION','DATE']])
    df_final = df_final.drop_duplicates(['STATION','DATE'])

    # Se pega la estacion centroide calculada al realizar la clusterizacion de estaciones
    df_final = df_final.merge(cat, on='STATION',how='left')
    
    # Se itera para el resto de variable del  datafreame en el que estamos intereasados
    for medida in ['TMAX','TMIN','TAVG','PRCP']:
        temporal = df[df.MEASURE == medida]
        temporal.VALUE = temporal.VALUE / 10.0
        temporal = temporal[["STATION","DATE","VALUE"]]
        temporal.columns = ["STATION","DATE",medida]
        df_final = pd.merge(df_final, temporal, how='outer', on=['STATION', 'DATE'])
    
    print("Numero de registros antes de la reduccion: " + str(len(df_final)))
    
    # Se realiza el groupby por la nueva estacion
    df_final = df_final.groupby(["NEW_STATION", "DATE"]).mean()
    df_final = df_final.reset_index()
        
    print("Numero de registros tras la reduccion: " + str(len(df_final)))
        
    df_final.to_csv('D:/master/Master-in-Data-Science/Proyecto/Datos/'+str(year)+'.csv',index=False)       
    
    # Escritura de los datos en la BBDD    
    query=''' insert into world_time (STATION,DATE,TMAX,TMIN,TAVG,PRCP) values (?,?,?,?,?,?) '''
    conn.executemany(query, df_final.to_records(index=False))
    conn.commit()        

# Se añade la columna que contiene la identificacion de pais y se procede a alimentar la variable
#res = cursor.execute("alter table world_time add column COUNTRY TEXT;");
#res = cursor.execute("update world_time set COUNTRY = substr(STATION,1,2);")
#conn.commit()

# Se crean los indices necesarios
#res = cursor.execute("create index idx_country on world_time (COUNTRY);");
#res = cursor.execute("create index idx_station on world_time (STATION);");
#res = cursor.execute("create index idx_date on world_time (DATE);");
#conn.commit()

# La tabla resultante es muy grande y el tiempo consumido en hacer consultas y generación de índices es prácticamente
# imposible por ese motivo se rompe la tabla por cada uno de los continentes a los que se localiza cada estacion

# Se lee la tabla que contiene la agrupación de contienentes por pais
cat_country = pd.read_csv('../datos/ghcnd-countries.csv',sep=",")

# Se recuperan todos los contienentes
continents = cat_country.drop_duplicates('CONTINENT')[['CONTINENT']].as_matrix()

for continent in continents:
    print ("Inicio del continente: " + continent)
    # Se seleccionan todos los  paises del continente
    countries = str(cat_country[cat_country.CONTINENT == continent[0]][['COUNTRY']].as_matrix()).replace("]]","").replace("[","").replace("]\n ",",")
    c = continent[0].replace(" ","_").lower()
    
    print ("Lista de paises del continente: " + countries)
    # Eliminacion de las tabla si existe previamente
    res = cursor.execute("drop table if exists world_time_"+c)
    res = cursor.execute("create table world_time_"+c+ """ as 
                    select 
                     STATION,
                     DATE,
                     TMAX,
                     TMIN,
                     TAVG,
                     PRCP 
                    from world_time
                    where substr(STATION,1,2) in (""" + countries +")")
                         
    conn.commit()
    
    # Se añade la columna que contiene la identificacion de pais y se procede a alimentar la variable
    res = cursor.execute("alter table world_time_"+c+ " add column COUNTRY TEXT;");
    res = cursor.execute("update world_time_"+c+" set COUNTRY = substr(STATION,1,2);")
    conn.commit()

    # Se crean los indices necesarios
    res = cursor.execute("create index idx_country on world_time_"+c+" (COUNTRY);");
    res = cursor.execute("create index idx_station on world_time_"+c+" (STATION);");
    res = cursor.execute("create index idx_date on world_time_"+c+ " (DATE);");
    
    print ("Fin del continente: " + continent)
    
# Se elimina la tabla global
res = cursor.execute("drop table world_time;")
    
conn.close()
